Auth and Policies

+++ app/Http/Controllers/PostController.php
@@ -3,54 +3,74 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\Post\StoreRequest;
+use App\Http\Requests\Tag\IdArrayRequest;
 use App\Models\Category;
 use App\Models\Post;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
-    public function index()
+    public function index(): View
     {
-        $posts = Post::with('category', 'tags')->get();
+        $posts = Post::with('category', 'tags', 'user')->get();
 
         return view('posts.index', compact('posts'));
 
     }
 
-    public function create()
+    public function create(): View
     {
         return view('posts.create');
     }
 
-    public function store(StoreRequest $request)
+    public function store(StoreRequest $request, IdArrayRequest $tagsRequest): RedirectResponse
     {
         $post = Post::query()->create($request->validated());
+        $post->tags()->attach($tagsRequest->tags);
 
         return redirect(route('posts.show', compact('post')));
     }
 
-    public function show(Post $post)
+    public function show(Post $post): View
     {
-        $post->load('category', 'tags');
+        $post->load('category', 'tags', 'user');
         return view('posts.show', compact('post'));
     }
 
-    public function edit(Post $post)
+    public function edit(Post $post): View
     {
+        Gate::authorize('postOwner', $post);
+        $post->load('tags');
         return view('posts.edit', compact('post'));
     }
 
-    public function update(Post $post, StoreRequest $request)
+    public function update(Post $post, StoreRequest $request, IdArrayRequest $tagsRequest): RedirectResponse
     {
+        Gate::authorize('postOwner', $post);
         $post->update($request->validated());
+
+        $post->tags()->sync($tagsRequest->tags);
 
         return redirect(route('posts.index'));
     }
 
-    public function destroy(Post $post)
+    public function destroy(Post $post): RedirectResponse
     {
+        Gate::authorize('postOwner', $post);
         $post->delete();
 
         return redirect(route('posts.index'));
+    }
+
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth', except: ['index']),
+        ];
     }
 }
